+++ tests/testthat/test-find.R
@@ -1,5 +1,38 @@
-context("test-find")
+test_that("primitive functions are never supersets", {
+  pkgload::load_all(test_path("primitive"), quiet = TRUE)
+  on.exit(pkgload::unload("primitive"))
 
+  expect_false(is_superset("sum", "primitive", "base"))
+  expect_equal(
+    superset_principle("sum", c("primitive", "base")),
+    c("primitive", "base")
+  )
+})
+
+test_that("superset", {
+  # by definition/design, there are no real conflicts in base functions
+  expect_equal(superset_principle("cbind", c("base", "methods")), character())
+
+  # Automatically created S4 generics obey the superset principle
+  expect_equal(superset_principle("print", c("base", "Matrix")), character())
+  # Even if the arguments have been customised
+  expect_equal(superset_principle("rcond", c("base", "Matrix")), character())
+})
+
+test_that("functions aren't conflicts with non-functions", {
+  pkgload::load_all(test_path("funmatch"), quiet = TRUE)
+  on.exit(pkgload::unload("funmatch"))
+
+  expect_equal(function_lookup("pi", c("base", "funmatch")), character())
+  expect_equal(function_lookup("mean", c("base", "funmatch")), character())
+})
+
+test_that("can find conflicts with data", {
+  pkgload::load_all(test_path("data"), quiet = TRUE)
+  on.exit(pkgload::unload("data"))
+
+  expect_named(conflict_scout(c("datasets", "data")), "mtcars")
+})
 
 # moved functions ----------------------------------------------------