+++ tests/testthat/test-zzz.R
@@ -1,10 +1,9 @@
-context("test-zzz.R")
-
-test_that("detaching removes shims", {
-  # I honestly don't understand why this works - detaching the package
-  # being tested seems like a bad bad idea; maybe it's ok because it's the
-  # last test?
-  testthat::expect_true(conflicts_attached())
-  detach(package:conflicted)
-  testthat::expect_false(".conflicts" %in% search())
+test_that("unloading removes shims", {
+  # Do in separate process to avoid interferring with tests
+  in_search <- callr::r(function() {
+    library(conflicted)
+    pkgload::unload("conflicted")
+    ".conflicts" %in% search()
+  })
+  expect_false(in_search)
 })