+++ NEWS.md
@@ -1,3 +1,23 @@
+# conflicted 1.1.0
+
+* New `conflicted_prefer_all()` and `conflicted_prefer_matching()` to
+  prefer functions en masse (#51).
+
+* Improvements to conflict detection and resolution:
+
+  * Reports conflicts involving lazy loaded datasets (#54).
+  
+  * Don't report conflicts involving a `standardGeneric` (#47).
+  
+  * Better handling of conflicts cleared by superset principle: if there is
+    a conflict all functions (including any base functions) are reported, and
+    if there isn't a conflict, no packages are reported (instead of 1) (#47).
+  
+  * Don't report conflict between a function and a non-function (#30).
+
+  * Conflicts involving a primitive function no longer error 
+    (@nerskin, #46, #48).
+
 # conflicted 1.0.4
 
 * Fixes for dev rlang